Integer Conversion Issues from GCC

Commit migrated from https://github.com/dotnet/coreclr/commit/837ac49c1ac46729df88ddb4a5d73ade6cc55cfc

+++ src/clr/src/ToolBox/SOS/lldbplugin/setsostidcommand.cpp
@@ -23,7 +23,7 @@ public:
     {
         if (arguments[0] == NULL)
         {
-            if (g_currentThreadSystemId == -1 || g_currentThreadIndex == -1)
+            if (g_currentThreadSystemId == (ULONG)-1 || g_currentThreadIndex == (ULONG)-1)
             {
                 result.Printf("sos OS tid not mapped\n");
             }
@@ -33,8 +33,8 @@ public:
             }
         }
         else if (strcmp(arguments[0], "-clear") == 0) {
-            g_currentThreadIndex = -1;
-            g_currentThreadSystemId = -1;
+            g_currentThreadIndex = (ULONG)-1;
+            g_currentThreadSystemId = (ULONG)-1;
             result.Printf("Cleared sos OS tid/index\n");
         }
         else if (arguments[1] == NULL)
